Introduce concept of "Serial sets" -- these are "virtual sets" that have real sets in them. When building all, a serial set builds it contents in the order declared. This allows for interset build dependencies (See perl yamlspecs/packages.yaml for an example).
Modify gen-definitions.py to add two query options:
--raw :  provide the python structure (e.g. a list with [ ]) instead of
just contents
--values: for 1-level dictionaries, just return the values of all
dictionary entries as a single list.

Add to run parallel the ability to take a template,pattern,and a list
(or dictionary) of values to create shell jobs that run in parallel.
If a dictionary, each dictionary entry is a
"serial set", If a list, then all jobs are parallel
Template structure greatly simplifies the buildall-parallel structure